requirements.txt
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.4
cycler==0.10.0
idna==3.2
kiwisolver==1.3.1
matplotlib==3.4.2
MouseInfo==0.1.3
numpy==1.21.1
opencv-contrib-python==3.4.2.17
opencv-python==4.5.1.48
pandas==1.3.1
Pillow==8.1.2
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.2
pypiwin32==223
PyRect==0.1.4
PyScreeze==0.1.26
pytesseract==0.3.3
python-dateutil==2.8.1
PyTweening==1.0.3
pytz==2021.1
pywin32==301
requests==2.26.0
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
urllib3==1.26.6
